clang-format
Ran Clang Format on the entire code base
This will likely not be the last time I do this, but it should be the most drastic

Sorting the #includes wrecked the code base. So I patched a lot of files to properly
foward declare classes and include exports.h and COmmonDataModel.h where it was needed
Code base compiled at 13 minutes in testing. Will watch the build machine to see
if it slowed us down to much