Transmucosal:  Model in place, initial testing in progress.
Added the transmucosal system of equations to Drugs.cpp.  Code
builds and runs through entire scenario in which Oral Transmucosal
action is used.  Fentanyl is appearing in plasma.
Added helper functions to SESubstanceOralDose to check which route
(transmucosal or GI) is being used.
Added probes to Drugs file to track transmucosal state variables.