Transmucosal:  Reduced compartment and transit model for GI
Defining the GI using multiple compartments with transit data
derived from literature.  This addition combined with much
improved literature estimate of effective permeability of GI
to fentanyl has greatly improved timing of GI absorption.

Changed partition coefficient for fentanyl in oral mucosa to
be the same as that for skin calculated by BioGears (similar
properties).  This value gave better results for fentanyl than
the lit empirical relationship used previously.  Slight changes
to buccal and sublingual thicknesses to get timing of dose right.
Adding in some features of a GI absorption model that should
not change baselines for fentanyl too much but will help in the
future with oral antiobitic administration.