fix:mmcdaniel-fentanyl:Remove check preventing respiration rate < 4/min
Delete check that set tidal volume, respiration rate, and ventilation to
0 if length of respiration period was longer than 0.25 min.  There is
already a check for target tidal volume = 0.0, so we will not run in to
any divide by 0 errors without this check.  RR < 4 should be supported
and should self-police itself.  If you leave it there that long, patient
will enter irreversible hypoxia, as would be expected.  Re-ran anesthesia machine scenarios and they were not affected by this change.