feature(cmd_bio) New GENSEPSIS option for generating sepsis states
      The build machine needs to be able to generate long term states in
the background. So we made a new GenSepsisstates.config file last commit
and now the cmd_bio util can load it if passed GENSEPSIS. Also patched
in the ability to load EngineStates and Inline Patients which isn't
fully tested and could be cleaned up a little as the first round was in
a rush.