f/mmcdaniel-baroreceptors: Tune nervous model to patient actions
CNS drug modifier was being applied to output of chemoreceptor function
(total ventilation target) but not to the chemoreceptor signal itself.
The signal kept increasing during fentanyl and morphine scenarios, which
caused improper downstream effects in Central Signal and Efferent
response functions.

Suppress chemoreceptor signal when sedatives and neuromuscular blockers
are applied.  This prevents the signal from accumulating when patient
intubated.

Set up anesthetics to change the sensitivity (slope response) of
baroreceptors rather than the barorecetor operating point.  This is more
in line with literature for drugs like propofol.