f/mmcdaniel-respDriver: Decrease rapid gains in respiratory circuit
Airway Obstruction and Bronchoconstriction action introduce large
resistance gains in one time step that cause respiratory circuit
instabilities.  This error was most noticeable when removing an
obstruction/constriction action of max severity and returning to
baseline.  Add a damping factor that gradually returns to baseline so
that circuit remains stable.