f/mmcdaniel-respDriver: Align anesthesia machine mask logic with driver
Set up the anesthesia machine mask setting so that its inahle/exhale
logic follows that of the respiratory driver.  Previously, a patient who
was not given a neuromuscular blocker could have the mask applied and
would not be able to breath.  This would occur because the respiratory
drive could be on an inhale phase while the anesthesia machine was on an
exhale phase (in which case the inspiratory valve is shut).

We will probably need separate logic if a patient is given a
neuromuscular blocker and a mask is applied.  However, I would think
that the most frequent applications would involve either mask combined
with spontaneous breathing or neuromuscular blocker combined with
intubation.

Updating anesthesia machine scenarios so that anything with sux
(neuromuscular blocker) uses intubation setting.  Also made slight
adjustment to mask setting for Desflurane.