f/mmcdaniel-baroreceptors: Introduce muscle and heart autoregulation
Muscle and heart resistances subject to autoregulation as function of
local oxygen levels.  This method helps capture drop in total peripheral
resistance that occurs during hypoxia in spite of increase in extrasplanchnic
resistance.

Removed resistance region labels as it is easy to keep track of muscle
and heart separately without them.

Testing with inactive atrial receptors--these will probably need to be
removed.