f/sawhite-logger - Modified Logger inheritance model and took out concrete interface calls.
While working on the UI and the UE4Plugin code I noticed some side effects in our Logger ownership.
BioGears would remove the LogFoward on distruction regardless of if it owned the log. I concider this a
major bug that is the oppisite of the purpose of passing a Logger.

While updating the code I simplified our Logger hiarchy and removed unused logging implementations.

As part of this change I removed a call to .str("") of the passed stream. Because of the way stream works. Most developers might expect there stream to be cleared while others may not
I do not know what the best behavior is.