f/sawhite-cmake - Improved INSTALL logic to support Biogears_BREAKOUT_INSTALL and Biogears_PLATFORM_PREFIX
The UE4Plugin requires installation code to be fanned out based on build variant. With this change you can use

BioGears_BREAKOUT_INSTALL to toggle this functionality. It is a hidden booelan variable that is false by default.
BioGears_PLATFORM_PREFIX  to add an additional fanout to the breakout installation

The installation structure will now look like the following if both variables are used.

<CMAKE_INSTALL_PREFIX>/lib
   `-msvc16/
     |---Release
     `---Debug
   `-android19/
     |---Release
     `---Debug