f/sawhite-stateload - :bug: BioGears now calls SetUp durring a State Load this allows a physiology engine to be reused across multiple state jumps.
This issue was affecting multiple integration events. Instead of being able to load multiple states in a game based on timejumps
A integrator would need to destroy and recreate an engine with each state load. This implies that LoadState should not be its own function
and instead should be a constructor paramater.  This fix will reinitialize the patient on state_load and hopefully avoid this problem.
Once conern is that this will invalidate SubstanceMgr which will cause DataTracker and ScenarioExec issues