f/sawhite-datatracks - libbiogears - Allow user to request appending data tracks instead of truncating output files. THis is for downstream integrators who may wish to do a lot of state loading, but with the same output file. I'm going to try and expose reprinting headers to the output file post state load to create timeline marks if thats important to the end user
Changes the behavior of load state which now retains all previous data
request from the last state.  I do not believe we aquire request stored
in the state, but if we do this behavior will be removed in a future
patch.

Purpose: State loading should be universal regardless of the state
creators orignal data request. Allowing them to be used in state graph
based applications with out complicating logging.

Users can choose to change anything about the DataRequest after a state
load, but the default should not require the user to act.