f/sawhite-substance Updated HowTo-API to deal with review comments.  I've noticed how fragil our DataTrack setup is and I suggest we modify  it
The problem can easily be seen in the following example

_pimpl->engine->GetEngineTrack()->GetDataRequestManager().CreateSubstanceDataRequest().Set(*_pimpl->engine->GetSubstanceManager().GetSubstance(component.GetSubstance().GetName()),
"PlasmaConcentration", MassPerVolumeUnit::ug_Per_L);

The engineTrack is created as an empty DataTrack which causes issues
when CreateSubstanceDataRequest().  And Then Set which can return false
stores internal variables only if the input user string exist.

We should depricate SET and make CreateSubstanceDataRequest take in
paramaters. It can then either return error codes or throw based on the
build configuration but will not actually create the request unless it
actually exist