Swig Bindings REV 2.0
Working on improving the SWIG bindings for readability. In the middle of the refactor
I noticed that the improvments made for CSHARP break the bindings for PYTHON this is unfortunate, but can be worked around
HOwever, I'm currently only funded for CSHARP so I'll go back and wrap the module import statements in #ifdefs to
include them by header file with module name for python but by import module.i for CSharp.